% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{generate_graphs}
\alias{generate_graphs}
\title{Generate graphs for kelp forest monitoring}
\usage{
generate_graphs(inDF, inLevels, prefix)
}
\arguments{
\item{inDF}{input data frame}

\item{inLevels}{input levels}

\item{prefix}{prefix for output figures}
}
\value{
NULL generates output figures as jpg
}
\description{
Generate graphs for kelp forest monitoring
}
