% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{get_dir_tables}
\alias{get_dir_tables}
\title{Get directory of tables with CSV's for R}
\usage{
get_dir_tables(nps_config)
}
\arguments{
\item{nps_config}{NPS configuration list object}
}
\value{
path to directory of tables with CSV's for R. Evaluates the machine
  name that my be inserted into the NPS configuration file, per: \code{Sys.info()[["nodename"]]}
}
\description{
Get directory of tables with CSV's for R
}
\examples{
nps_config <- read_yaml(here("data/nps_config.yaml"))
get_dir_tables(nps_config)
}
