% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{get_n_spp_pivtbl}
\alias{get_n_spp_pivtbl}
\title{Get pivot table of species richness by park and year}
\usage{
get_n_spp_pivtbl(cfg, park, year, xlsx = NULL)
}
\arguments{
\item{cfg}{NPS configuration list object; see \code{\link{get_nps_config}}}

\item{park}{NPS park abbreviation}

\item{year}{year of data to extract}

\item{xlsx}{optional Excel spreadsheet path to write out}
}
\value{
pivottabler::PivotTable R6 object. Render with n_spp_tbl$renderPivot()
}
\description{
Get pivot table of species richness by park and year
}
\examples{
cfg  <- get_nps_config(system.file(package="npstools", "nps_config.yaml"))
park <- "CABR"
year <- 2015

n_spp_pivtbl <- get_n_spp_pivtbl(cfg, park, year)

# render pivot table as html widget
n_spp_pivtbl$renderPivot()
}
