% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{get_nps_config}
\alias{get_nps_config}
\title{Read NPS configuration file}
\usage{
get_nps_config(nps_config_yaml)
}
\arguments{
\item{nps_config_yaml}{NPS configuration file in YAML format}
}
\value{
list object from reading in the NPS configuration file
}
\description{
Read NPS configuration file
}
\examples{
get_nps_config(system.file(package="npstools", "nps_config.yaml"))
}
