% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{get_pct_cover_tbl}
\alias{get_pct_cover_tbl}
\title{Get table of absolute percent cover for given park and year}
\usage{
get_pct_cover_tbl(cfg, park, year)
}
\arguments{
\item{cfg}{NPS configuration list object; see \code{\link{get_nps_config}}}

\item{park}{park abbreviation, eg "CABR", "CHIS" or "SAMO"}

\item{year}{4-digit year}
}
\value{
Tibble that reproduces from \href{
      https://github.com/ecoquants/npstools/blob/3ca70ac9704a4a11d6d5d34f707e3008e35d0a35/inst/accdb_source/mod_ExportQueries.vb}{
      mod_ExportQueries}:

\itemize{
  \item{
    \href{
      https://github.com/ecoquants/npstools/blob/3ca70ac9704a4a11d6d5d34f707e3008e35d0a35/inst/accdb_source/mod_ExportQueries.vb#L1225-L1289}{
      Export_AnnualReport_AbsoluteCover()}}}

for "Figure E.2. Absolute foliar cover (\%) of plant growth forms, as observed during 20XX monitoring at CABR. Colored bars show mean values, while error bars extend ±1 s.d. from the means." from MEDN_veg_protocol_NARRATIVE_FINAL_8Sep2016.pdf.
}
\description{
Get table of absolute percent cover for given park and year
}
\examples{
cfg  <- get_nps_config(system.file(package="npstools", "nps_config.yaml"))
park <- "CABR"
year <- 2015

get_pct_cover_tbl(cfg, park, year)
}
