% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{get_spp_richness_plots}
\alias{get_spp_richness_plots}
\title{Get species richness plots}
\usage{
get_spp_richness_plots(park, year)
}
\arguments{
\item{park}{not yet implemented}

\item{year}{not yet implemented}
}
\value{
tibble
}
\description{
Get species richness plots
}
