% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{get_total_eventpoints_tbl}
\alias{get_total_eventpoints_tbl}
\title{Get table of total event points for given park}
\usage{
get_total_eventpoints_tbl(park)
}
\arguments{
\item{park}{park abbreviation, eg "CABR", "CHIS" or "SAMO"}
}
\description{
Get table of total event points for given park
}
