% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{read_park_tables}
\alias{read_park_tables}
\title{Read Park Tables}
\usage{
read_park_tables(nps_config, park, append_park = F)
}
\arguments{
\item{nps_config}{configuration of paths, etc}

\item{park}{park code (ie CHIS, CABR or SAMO)}

\item{append_park}{whether to append park code to name (eg tbl_Species_CHIS)}
}
\value{
Does not return anything. Loads all tables listed in the nps_config$dir_tables/park folder into the global namespace.
}
\description{
Read Park Tables
}
