% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{get_n_spp_pivtbl}
\alias{get_n_spp_pivtbl}
\title{Get pivot table of species richness by park and year}
\usage{
get_n_spp_pivtbl(park, year, xlsx = NULL)
}
\arguments{
\item{park}{NPS park abbreviation}

\item{year}{year of data to extract}

\item{xlsx}{optional Excel spreadsheet path to write out}
}
\value{
pivottabler::PivotTable R6 object. Render with n_spp_tbl$renderPivot()
}
\description{
Get pivot table of species richness by park and year
}
\examples{
nps_config <- get_nps_config(here("data/nps_config.yaml"))
park <- "CINMS"
year <- 2015

# optional path to Excel spreadsheet output
n_spp_xlsx <- here(glue("data/spp_richness_pivot_{park}_{year}.xlsx"))

n_spp_tbl <- get_n_spp_tbl(park, year, xlsx=n_spp_xlsx)

# render pivot table as html widget
n_spp_tbl$renderPivot()
}
