% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{get_spp_park_tbl}
\alias{get_spp_park_tbl}
\title{Get table of species data for given park}
\usage{
get_spp_park_tbl(park)
}
\arguments{
\item{park}{park abbreviation, eg "CABR", "CHIS" or "SAMO"}
}
\value{
tibble
}
\description{
Get table of species data for given park
}
\examples{
nps_config <- get_nps_config(here("data/nps_config.yaml"))
get_spp_park_tbl(park = "CABR")
}
