% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{get_total_eventpoints_tbl}
\alias{get_total_eventpoints_tbl}
\title{Get table of total event points for given park}
\usage{
get_total_eventpoints_tbl(cfg, park, reload = T)
}
\arguments{
\item{cfg}{NPS configuration list object; see \code{\link{get_nps_config}}}

\item{park}{park abbreviation, eg "CABR", "CHIS" or "SAMO"}
}
\value{
tibble with fields: Park, IslandCode, Location_ID, SiteCode,
  Vegetation_Community, SurveyYear, SurveyDate, NofPoints
}
\description{
Get table of total event points for given park
}
\examples{
cfg <- get_nps_config(system.file(package="npstools", "nps_config.yaml"))
get_total_eventpoints_tbl(cfg, park)
}
