% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_park_tables}
\alias{load_park_tables}
\title{Read Park Tables}
\usage{
load_park_tables(cfg, park, tbls = NULL, append_park = F)
}
\arguments{
\item{cfg}{NPS configuration list object; see \code{\link{get_nps_config}}}

\item{park}{park code (ie CHIS, CABR or SAMO)}

\item{tbls}{character vector of tables to load (default=NULL, loads all tables)}

\item{append_park}{whether to append park code to name (eg tbl_Species_CHIS)}
}
\value{
Does not return anything. Loads all tables listed in the cfg$dir_tables/park folder into the global namespace.
}
\description{
Read Park Tables
}
\examples{
cfg <- get_nps_config(system.file(package="npstools", "nps_config.yaml"))
load_park_tables(cfg, "CABR", tbls=c("tbl_Phenology_Species", "tlu_Richness"))

}
